Fix `Window::as_x11:handle` (by @synchis)
This was not found by me, but was found by @\synchis from the Rust Audio discord (@\synchis
pls lmk who/where to credit). I couldn't find a PR fixing this, so I decided to make one (as
I was already mucking about in the extension code anyways). If the PR shows up, feel free to
close this one.